requirements.txt
langgraph>=0.0.26
langchain-openai>=0.0.8
langchain-core>=0.1.28
python-dotenv>=1.0.1
pexpect>=4.8.0
playwright>=1.31.0
pydantic>=2.0.0
